nexus_ai/requirements.txt
streamlit>=1.38.0
google-genai>=1.0.0
supabase>=2.0.0
python-docx>=1.1.0
